requirements/production.txt
-r ./base.txt
daphne==3.0.2
psycopg2==2.9.3
channels-redis==3.3.1
celery==5.2.3
redis==4.1.0
